Add Mobile Interface for 3D view
Add an interface to consult 3D models with a mobile.
You can move in the 4 directions, turn left and right the camera, and go to previous / next design.

UX Improvement : change brightness of buttons on mouseover and on click, in order to show the user that they work.
Other : Add video in dummy offers page.