Improvement MTL changer Python
Improve the MTL changer Python File in order to remove the materials without file before ThreeJS, so in material file
Minor fixes : in univers.php, when click on a map, we have a new page and a link to the 3D model of this map.
The music is now disable when we change the map.