Update when logs file is loaded
Now, jdrCombatLogs is loaded only when we are going to save. So it is loaded 2 times (for turn check and end of fight). The ID is generated only one time to avoid to increment (because someone do a fight) between "log" and "end of fight".

New Zumis monsters icons.
Fix margin to see "Afficher le site" at the bottom.

In eqpt, remove "reduction dégât phys/mag". Update regex "PV +X"
+ 14 eqpts